Split extension manifest surface row
