+++ Tests/QuillCodeParityTests/ParityGateTests.swift
@@ -1900,14 +1900,18 @@ final class ParityGateTests: XCTestCase {
     func testWorkspaceSurfaceDelegatesSecondaryPaneSurfaceContracts() throws {
         let surfaceText = try Self.appSourceText(named: "WorkspaceSurface.swift")
         let secondaryText = try Self.appSourceText(named: "QuillCodeSecondaryPaneSurface.swift")
+        let extensionRowText = try Self.appSourceText(named: "ProjectExtensionManifestSurface.swift")
 
         XCTAssertTrue(secondaryText.contains("public struct WorkspaceExtensionsSurface"), "Extensions surface should live beside secondary-pane contracts.")
         XCTAssertTrue(secondaryText.contains("public struct WorkspaceMemoriesSurface"), "Memories surface should live beside secondary-pane contracts.")
         XCTAssertTrue(secondaryText.contains("public struct WorkspaceAutomationsSurface"), "Automations surface should live beside secondary-pane contracts.")
-        XCTAssertTrue(secondaryText.contains("public struct ProjectExtensionManifestSurface"), "Extension manifest rows should live beside secondary-pane contracts.")
         XCTAssertTrue(secondaryText.contains("public struct MemoryNoteSurface"), "Memory note rows should live beside secondary-pane contracts.")
         XCTAssertTrue(secondaryText.contains("public struct AutomationWorkflowSurface"), "Automation workflow rows should live beside secondary-pane contracts.")
-        XCTAssertTrue(secondaryText.contains("MCPToolDescriptor"), "MCP probe display compatibility should stay with extension surface rows.")
+        XCTAssertTrue(secondaryText.contains("ProjectExtensionManifestSurface("), "Extensions surface should still delegate row projection to extension manifest rows.")
+        XCTAssertTrue(extensionRowText.contains("public struct ProjectExtensionManifestSurface"), "Extension manifest rows should live in a focused surface row file.")
+        XCTAssertTrue(extensionRowText.contains("MCPToolDescriptor"), "MCP probe display compatibility should stay with extension surface rows.")
+        XCTAssertTrue(extensionRowText.contains("public init(from decoder: Decoder)"), "Extension row decode compatibility should stay with the row contract.")
+        XCTAssertFalse(secondaryText.contains("public struct ProjectExtensionManifestSurface"), "Secondary pane aggregate should not own extension manifest row internals.")
         XCTAssertFalse(surfaceText.contains("public struct WorkspaceExtensionsSurface"), "WorkspaceSurface should not own Extensions surface records.")
         XCTAssertFalse(surfaceText.contains("public struct WorkspaceMemoriesSurface"), "WorkspaceSurface should not own Memories surface records.")
         XCTAssertFalse(surfaceText.contains("public struct WorkspaceAutomationsSurface"), "WorkspaceSurface should not own Automations surface records.")
